Fixes the integration test
During the integration test, we're cloning the SRIOV Network Device
Plugin repository in order to deploy it.

However, it currently fails in the CI because we're cloning it in the
test instance, rather than cloning it locally. This results in
FileNotFoundErrors because the files we're trying to read to not exist
locally.

This updates the integration test to clone the git repository locally,
rather than in the test instance.